Create CapitaliseFirstLettrOfNames.js
I had this question at an interview recently.  
You have an array of strings of peoples names but they are capitalized incorrectly.  You need to regularize the sting and remove unnecessary whitespace.  
Finally convert the resulting array into a string where commas separate the elements. 
If the array is empty, return an empty string.